requirements.txt
# OCI Delivery Agent Function Dependencies
# Core dependencies
langchain==0.1.0
pillow>=9.0.0
python-dotenv>=0.19.0
oci>=2.0.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# numpy>=1.21.0
# pandas>=1.3.0
# requests>=2.25.0
